requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.2
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1
django-crispy-forms==1.9.2
django-filter==2.3.0
django-flatpickr==1.0.1
django-js-asset==1.2.2
django-phonenumber-field==4.0.0
django-summernote==0.8.11.6
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.49
Pillow==9.1.1
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.5.4
urllib3==1.25.9
wrapt==1.12.1
gunicorn
psycopg2>=2.1
